Added warning message to nvidia_smi xmldevicequery
I think this was only ever meant to be an example in the original nvml
library. Given that nvidia update the nvidia smi query, it will be too
hard to keep this query up to date. Instead, calling 'nvidia-smi -q -x'
from python with subprocess would achieve the same result.